ADD ORDERS REPORT

+++ src/handlers/callbackHandler.ts
@@ -4,7 +4,7 @@ import { redis, rStates, ttls } from "../redis";
 import { handleStartMenu } from "../components/botAnswers";
 import { RediceService } from "../bot";
 import { createEditData, MessageService } from "../services/messageManipulatorService";
-import { articleOptions, CallbackData, generateArticlesButtons, generateReportTimeButtons, mainOptions, Options, returnMenu, yesNo } from "../components/buttons";
+import { articleOptions, CallbackData, generateArticlesButtons, generateReportTimeButtons, mainOptions, Options, ordersReportMenu, returnMenu, yesNo } from "../components/buttons";
 import { users_db } from "../../database/models/users";
 import { getCurrentArticle, parseArticleData } from "../utils/parse";
 import { articles_db } from "../../database/models/articles";
@@ -249,23 +249,21 @@ export async function callbackHandler(query: TelegramBot.CallbackQuery, bot: Tel
       break;
 
     case 'orders report': 
-      // const accessOrdersReport = isReportAvailable(last_sec_report_call);
-      // if (accessStockReport) {
-        // await users_db.updateLastReportCall(chat_id, 'last_sec_report_call');
-        MS.deleteAllMessages(chat_id);
-        const loadingMsg = await bot.sendMessage(chat_id, texts.loadingReports, { disable_notification: true });
-    
-        try {
-          requestOrdersReport(chat_id, loadingMsg.message_id); 
-        } catch (error) {
-          console.error("Failed to generate report:", error);
-          if (mainBtn) editData = createEditData(texts.errorGetSkuAgain, mainBtn);
+        const ordersCallback = userCallbackData.split('?')
+
+        if (!ordersCallback[1]) {
+          editData = createEditData(texts.errorGetSkuAgain, ordersReportMenu());
+        } else {
+          MS.deleteAllMessages(chat_id);
+          const loadingMsg = await bot.sendMessage(chat_id, texts.loadingReports, { disable_notification: true });
+      
+          try {
+            requestOrdersReport(chat_id, loadingMsg.message_id, ordersCallback[1] as "today" | "yesterday"); 
+          } catch (error) {
+            console.error("Failed to generate report:", error);
+            if (mainBtn) editData = createEditData(texts.errorGetSkuAgain, mainBtn);
+          }
         }
-    
-        // await MS.deleteMessage(chat_id, loadingMsg.message_id);
-      // } else {
-      //   if (mainBtn) editData = createEditData(texts.errorGetSkuAgain, mainBtn);
-      // }
       break;
 
     case 'ref':
